+++ Makefile
@@ -1,0 +1,105 @@
+#* Variables
+SHELL := /usr/bin/env bash
+PYTHON := python
+PYTHONPATH := `pwd`
+
+#* Docker variables
+IMAGE := vauth
+VERSION := latest
+
+#* Poetry
+.PHONY: poetry-download
+poetry-download:
+	curl -sSL https://install.python-poetry.org | $(PYTHON) -
+
+.PHONY: poetry-remove
+poetry-remove:
+	curl -sSL https://install.python-poetry.org | $(PYTHON) - --uninstall
+
+#* Installation
+.PHONY: install
+install:
+	poetry lock -n && poetry export --without-hashes > requirements.txt
+	poetry install -n
+	-poetry run mypy --install-types --non-interactive ./
+
+.PHONY: pre-commit-install
+pre-commit-install:
+	poetry run pre-commit install
+
+#* Formatters
+.PHONY: codestyle
+codestyle:
+	poetry run pyupgrade --exit-zero-even-if-changed --py39-plus **/*.py
+
+.PHONY: formatting
+formatting: codestyle
+
+#* Linting
+.PHONY: test
+test:
+	PYTHONPATH=$(PYTHONPATH) poetry run pytest -c pyproject.toml --cov-report=html --cov=vauth tests/
+
+.PHONY: check-codestyle
+check-codestyle:
+	poetry run darglint --verbosity 2 vauth tests
+	poetry run ruff --diff --config pyproject.toml ./
+	poetry run yamllint -c .yamllint ./
+
+.PHONY: mypy
+mypy:
+	poetry run mypy --config-file pyproject.toml ./
+
+.PHONY: lint
+lint: test check-codestyle mypy check-safety
+
+# "isort[colors]@latest"
+# poetry add -D --allow-prereleases black@latest
+.PHONY: update-dev-deps
+update-dev-deps:
+	poetry add -D darglint@latest ruff@latest yamllint@latest mypy@latest pre-commit@latest pydocstyle@latest pylint@latest pytest@latest pyupgrade@latest safety@latest pytest-html@latest pytest-cov@latest
+	
+#* Docker
+# Example: make docker-build VERSION=latest
+# Example: make docker-build IMAGE=some_name VERSION=0.1.0
+.PHONY: docker-build
+docker-build:
+	@echo Building docker $(IMAGE):$(VERSION) ...
+	docker build \
+		-t $(IMAGE):$(VERSION) . \
+		-f ./.devcontainer/Dockerfile --no-cache
+
+# Example: make docker-remove VERSION=latest
+# Example: make docker-remove IMAGE=some_name VERSION=0.1.0
+.PHONY: docker-remove
+docker-remove:
+	@echo Removing docker $(IMAGE):$(VERSION) ...
+	docker rmi -f $(IMAGE):$(VERSION)
+
+#* Cleaning
+.PHONY: pycache-remove
+pycache-remove:
+	find . | grep -E "(__pycache__|\.pyc|\.pyo$$)" | xargs rm -rf
+
+.PHONY: dsstore-remove
+dsstore-remove:
+	find . | grep -E ".DS_Store" | xargs rm -rf
+
+.PHONY: mypycache-remove
+mypycache-remove:
+	find . | grep -E ".mypy_cache" | xargs rm -rf
+
+.PHONY: ipynbcheckpoints-remove
+ipynbcheckpoints-remove:
+	find . | grep -E ".ipynb_checkpoints" | xargs rm -rf
+
+.PHONY: pytestcache-remove
+pytestcache-remove:
+	find . | grep -E ".pytest_cache" | xargs rm -rf
+
+.PHONY: build-remove
+build-remove:
+	rm -rf build/
+
+.PHONY: cleanup
+cleanup: pycache-remove dsstore-remove mypycache-remove ipynbcheckpoints-remove pytestcache-remove
