+++ requirements.txt
@@ -1,0 +1,23 @@
+annotated-types==0.6.0 ; python_version >= "3.10" and python_version < "3.12"
+anyio==3.7.1 ; python_version >= "3.10" and python_version < "3.12"
+async-timeout==4.0.3 ; python_version >= "3.10" and python_full_version <= "3.11.2"
+cffi==1.16.0 ; python_version >= "3.10" and python_version < "3.12"
+click==8.1.7 ; python_version >= "3.10" and python_version < "3.12"
+colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.12" and platform_system == "Windows"
+cryptography==41.0.7 ; python_version >= "3.10" and python_version < "3.12"
+exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
+fastapi==0.103.2 ; python_version >= "3.10" and python_version < "3.12"
+hiredis==2.2.3 ; python_version >= "3.10" and python_version < "3.12"
+idna==3.6 ; python_version >= "3.10" and python_version < "3.12"
+more-itertools==9.1.0 ; python_version >= "3.10" and python_version < "3.12"
+pycparser==2.21 ; python_version >= "3.10" and python_version < "3.12"
+pydantic-core==2.3.0 ; python_version >= "3.10" and python_version < "3.12"
+pydantic==2.0.3 ; python_version >= "3.10" and python_version < "3.12"
+python-ulid==1.1.0 ; python_version >= "3.10" and python_version < "3.12"
+redis-om==0.2.1 ; python_version >= "3.10" and python_version < "3.12"
+redis==4.6.0 ; python_version >= "3.10" and python_version < "3.12"
+sniffio==1.3.0 ; python_version >= "3.10" and python_version < "3.12"
+starlette==0.27.0 ; python_version >= "3.10" and python_version < "3.12"
+types-pyopenssl==23.3.0.0 ; python_version >= "3.10" and python_version < "3.12"
+types-redis==4.6.0.11 ; python_version >= "3.10" and python_version < "3.12"
+typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "3.12"
